mcp_server Phase 3: split the 21 tools into tools/*.py by topic
Final phase of the tools/mcp_server/ modularization (prog.txt Sezione 3).
server.py drops from ~700 lines to a thin entrypoint (9 lines of actual
logic): creates the MCPServer instance, imports each tools/*.py module so
its @mcp.tool decorators register against it, re-exports every tool
function (so `from mcp_server.server import dense_evolution_health` keeps
working for existing callers, including this repo's own test suite), and
provides main().

- tools/system_tools.py: health, system_limits, list_presets, list_gates,
  list_noise_models, list_molecules
- tools/circuit_tools.py: build_circuit, run_circuit
- tools/chemistry_tools.py: molecule_energy, mix_molecules,
  custom_molecule_energy, energy_scan, run_vqe, qmmm_forces, md_trajectory
- tools/mitigation_tools.py: mitigate_zne, mitigate_density_matrix,
  vector_healing
- tools/wormhole_tools.py: wormhole_select_instance,
  wormhole_teleportation, wormhole_scan

New molecules.py holds molecule-catalog resolution (_resolve_molecule_name,
_get_annotated_molecule_catalog, the TTLCache instance) -- shared by
system_tools.py and chemistry_tools.py, so it doesn't belong to either one
specifically.

Each tools/*.py module does `from ..server import mcp` -- looks circular
but is safe: mcp = MCPServer(...) is assigned in server.py BEFORE its
`from .tools.X import ...` lines run, so by the time each submodule's
import executes, mcp_server.server is already in sys.modules with `mcp`
already set. Verified directly, not just reasoned about: all 21 tools
import and register correctly.

Test suite updated for two more Phase-1/2-style relocations: Input model
classes (`mcp_adapter.XInput(...)` -> `mcp_models.XInput(...)`, since
server.py no longer re-exports them just for test convenience) and the
molecule catalog cache (`mcp_adapter._molecule_catalog_cache` ->
`mcp_molecules._molecule_catalog_cache`). Also fixed a stale test-only
monkeypatch target left over from the Phase 2 coverage fix
(server_module._request -> mcp_molecules._request, since
_get_annotated_molecule_catalog now lives in molecules.py with its own
bound _request).

Also closed two more real coverage gaps found the same way as Phase 2
(local --cov-report=term-missing, not codecov's UI): `mcp.run()` marked
`# pragma: no cover` (blocks on the real stdio transport loop, genuinely
untestable), and a new test for energy_scan's real kernel-exception path
(the existing "isolates a failing point" test only exercised the LOCAL
symbols/geometry length check, never actually reached the try/except
around the real _request call -- a real gap in test coverage, not
something this refactor introduced, but now closed since it's visible in
the new file's patch coverage).

Verified: pyproject.toml's `packages` list updated with "mcp_server.tools"
(checked with a real `python -m build --wheel`, same lesson as Phase 2).
53/53 tests pass; tools/mcp_server is at 100% coverage locally.

Section 3 (dashboard MCP server refactor, prog.txt) is now complete.
